blockwise tuning

+++ include/QoZ/compressor/SZInterpolationCompressor.hpp
@@ -563,7 +563,7 @@ namespace QoZ {
     
                         std::array<size_t,N>sb_starts;
                         std::fill(sb_starts.begin(),sb_starts.end(),0);
-                        std::array<size_t,N>sb_ends;
+                        std::array<int,N>sb_ends;
                         std::fill(sb_starts.begin(),sb_starts.end(),-1);
                         
                         if(N==2){
